test: remove orphaned version sync module
