+++ src/tests/version-sync.ts
@@ -2,8 +2,6 @@ import { readFileSync } from 'node:fs';
 import { dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { extractBundleCardVersion } from '../utils/version-sync.js';
-
 interface CardVersionSnapshot {
   packageVersion: string;
   manifestVersion: string;
@@ -24,6 +22,23 @@ function parseVersionField(jsonText: string, sourceName: string): string {
   return parsed.version;
 }
 
+export function extractBundleCardVersion(bundleText: string): string {
+  const markerIndex = bundleText.indexOf('name:"AutoSnooze Card"');
+  if (markerIndex === -1) {
+    throw new Error('Could not locate AutoSnooze card metadata in the built bundle.');
+  }
+
+  const precedingSource = bundleText.slice(0, markerIndex);
+  const matches = [...precedingSource.matchAll(/function\s+\w+\(e="([^"]+)"\)/g)];
+  const version = matches.length > 0 ? matches[matches.length - 1]?.[1] : undefined;
+
+  if (!version) {
+    throw new Error('Could not extract the default registration version from the built bundle.');
+  }
+
+  return version;
+}
+
 export function readCardVersionSnapshot(): CardVersionSnapshot {
   return {
     packageVersion: parseVersionField(readFileSync(PACKAGE_JSON_PATH, 'utf8'), 'package.json'),
@@ -33,4 +48,18 @@ export function readCardVersionSnapshot(): CardVersionSnapshot {
     ),
     bundleVersion: extractBundleCardVersion(readFileSync(BUNDLE_PATH, 'utf8')),
   };
+}
+
+export function assertCardVersionSync(snapshot: CardVersionSnapshot): void {
+  if (snapshot.packageVersion !== snapshot.manifestVersion) {
+    throw new Error(
+      `manifest version "${snapshot.manifestVersion}" does not match package version "${snapshot.packageVersion}".`
+    );
+  }
+
+  if (snapshot.bundleVersion !== snapshot.packageVersion) {
+    throw new Error(
+      `bundle version "${snapshot.bundleVersion}" does not match package version "${snapshot.packageVersion}".`
+    );
+  }
 }
